Bug fix: version 1 ics writing
Because ICS writing depends on flags in the ICSFormat.Metadata class,
and these values were being directly copied from the Reader Metadata
(if reading and writing ICS) or not populated at all (if translating
to ICS), these flags were never correctly seti, causing the .ics file
not to be generated when writing version 1 ICS.

To fix this, the string-based setDest methods will now set the
Metadata's idsId appropriately. This will cause the accompanying .ics
file to be properly generated when writing version 1 ICS files.

Additionally an unnecessary null check was removed, as the object
(pixels) could only be null at that point.

Note that this solution puts an implicit requirement on calling the
String-based setters. This solution should either be migrated to the
RAOS-based signature, or the unsafe signatures should be marked
upsupported.

Further note that this solution is functional when provided a .ids dest,
but needs to be extrapolated for the .ics dest case.