Reworked AdapterTools mechanics
First, instead of maintaining a static function returning a queried
Adapter, a non-static class was created, and AdapterTools maintains a
static reference to a single instance of this helper class.

Second, instead of creating a list of adapters and only using what's on
it, if a queried adapter isn't found, a new instance is created and
mapped. This allows adapters to be defined downstream of AdapterTools
without any explicit discovery mechanism.