Refactored format name/suffixes to Format
Previously, the format file name and associated suffixes were being
tracked via a separate FormatHandler API, and included in several
components (e.g. Checkers, Writers) as needed.

This leads to repetition of code as each component is tracking the same
format. Instead it is logical that the actual Format interface be the
maintainer of format name and suffixes.

Thus all such references were migrated to the Format interface.

Static utility methods intended to operate on the format name and
suffixes were migrated to FormatTools.

All references and constructors were updated as appropriate.