Removed getZCTCoords from Reader API
The getZCTCoords method is, in general, a convenience for the static
FormatTools.getZCTCoords method and it seems unnecessary to make it a
part of the Reader API when it is typically just delegating to a utility
class.

If a format requires specific behavior for such a method, it can
maintain it within its own utility class.