Refactored static-context to be Legacy only
The single, static context located in ome.scifio FormatTools was
removed. Context can be created as needed for SCIFIO code.

In the legacy layer, context creation is modeled after ImageReader - in
which a list of readers is parsed once and statically cached. Using this
philosophy, a LegacyContext class was added, with a get() method to get
the current context.

The LegacyContext context (SCIFIO object) is created once and statically
cached.