Plane implementation
Added a Plane class hierarchy.

The Plane interface is untyped, with a generic plane type introduced by
DataPlane.

The two currently supported plane types are ByteArrayPlane and
BufferedImagePlane, which correspond to the performance of Bio-Formats
(returning a byte[] or BufferedImage).

Planes are intended to be returned by Readers, via openPlane methods,
and represent a plane of the Image currently being read.