Typed/Untyped interface split
For the base components (Reader, Parser, Metadata, Format, Translator,
Writer) that used generic typing, the interfaces have been split out so
that the lowest level API does not involve generics.

As concrete classes are able to specify what components they are related
to, the generics are rarely if ever present at the concrete level as
well.

This allows the user to be shielded from the complicated syntax of
generic parameters, while still gaining type safety in defining the
relationships between various components.

This commit includes the necessary refactoring for classes to use typed
or untyped components, as appropriate.