FakeFormat ColorTable available before openPlane
In SCIFIO, since planes contain the color table, no color table is
guaranteed to be available before openPlane is invoked.

This was true of bio-formats as well, but some formats (e.g. FakeFormat)
required planes before openBytes was called anyway.

To accommodate this in SCIFIO, specific Format implementations will have
to provide API for color table access.

Fixed in FakeFormat.