Updated scijava-common jar to fix NPE
Added a custom scijava-common jar that fixes an NPE when checking
versions from the pom.

Not sure why the NPE is thrown at all as the meta-inf looked good, but
this will be fixed before a release of scijava-common is cut.

This jar should be replaced with a release jar eventually.