getStream does not always log exceptions
The SCIFIOReaderWrapper getStream was almost always catching and
logging an exception. It now does a more thorough check to find the
input stream if available, and returns null if not, without throwing
an exception (unless there is a legitimate reason now).  # Please
enter the commit message for your changes.

This should help minimize log file sizes.