Updated ZipHandle constructors
Added a context constructor to the scifio ZipHandle, and changed the
legacy ZipHandle to use it appropriately.

The SCIFIO ZipHandle constructor was unnecessarily repeating itself, and
throwing NPEs because it would do things in the wrong order. This has
been fixed by delegating to the super (Context) constructor and simply
calling setFile.