Converted SCIFIO to context wrapper
Instead of making SCIFIO a service, so one would have to call
getContext.getService(SCIFIO.class), The SCIFIO class is now just a
disposable wrapper that can be constructed around an existing context to
provide convenience acccessors for SCIFIO-related services (which is all
it ever did, anyway).

This greatly simplifes the code, SCIFIOs can be constructed any time, as
needed.