Converted wrappers to SCIFIO filters
The wrappers previously located in ome/scifio/wrappers were
non-functional for SCIFIO.

A parallel set of wrappers (renamed filters) was designed to be
automatically discovered. The idea is that users can get a handle on a
top-level controller (e.g. ReaderFilter) which will automatically
discover and populate a list of filters, which can be enabled and
disabled at will (instead of necessitating a rebuiling of the filter
stack).

This makes wrapping in SCIFIO flexible and extensible.