Fixed BufferedImageReader's openThumbImage use
Two planes have to be created in this method - first a full size plane
is created and passed to AWTImageTools.openThumbImage. Then the
resultant plane is set as a thumb sized plane's BufferedImage data.

Previously, openThumbImage was being called on a the thumb plane itself,
which would be too small.