Added SCIFIOReader
Created a loci.formats SCIFIOReader. This reader was added to the bottom
of readers.txt, so it is checked last

It works like an ImageReader, capable of delegating to any SCIFIO Reader
implementation.

This allows new formats to be supported in SCIFIO that don't have a
parallel in Bio-Formats, but are still readable by the legacy code.