Removed WeakReference value from LegacyAdapters
Legacy adapters were using a weak hash map to map to WeakReference
values.

This was done because the values are wrappers of the keys, so having a
strong reference to the values creates a strong link back to the key, so
they would never be garbage collected.

However, there is no guarantee that the values will ever have a strong
reference to them. Given that the legacy code has no notion of the
wrappers, there would be no reason to have a reference to them.

Thus there must be some strong reference in existence, and so it makes
sense for it to be in these maps. The WeakReferences should be pushed
down to the individual wrappers, so that they wrap their targets
weakly.