Removed anonymous constructors from RAISAdapter
When constructing a RAIS wrapper from the adapter, a RAIS instance must
be provided, as anonymous construction will cause problems with
immediate garbage collection (and thus memory leaks).

Especially given that if the original stream is closed, even if the
wrapper RAIS is not gc'd, there's no way to close it.