Simplified LegacyAdapter API
It now follows a similar pattern as SCIFIO of having the low-level
interface be generic-free, but abstract and concrete layers specific
generic types.

This makes working with adapters easier, and facilitates a paradigm
where the user doesn't have to know what kind of adapter they are using.

Removal of types from the interface meant condensing the API to simple
"get" and "map" methods that operate on objects. Type checking is all
handled in the abstract layer, and by the static utility classes.

The one down side to these changes is that, since Objects are being
used, casting is required. To get around this, each package with
Adapters in it has a new static utility class with type-specific methods
dictating the relationship between legacy and adapter clases.