Moved Formats to ome.scifio.formats package
Formats were originally put in their own package because of the number
of classes involved in defining a format (e.g. each SCIFIO component,
plus translators and utility classes).

Now that all Format-related classes are implemented as nested classes it
is unnecessary to also put them in their own package. The "in" and "out"
package conventions of Bio-Formats are also not appropriate, since the
Format class can contain both readers and writers.

Thus "formats" seems an appropriate package name.