AVIFormat conversion
The loci.formats AVIReader and AVIWriter have been converted to a single
AVIFormat in ome.scifio.formats.

The legacy classes are new delegators, which pass all functionality to
the SCIFIO implementation.

Supporting classes have been copied to SCIFIO (codecs and the
UnsupportedCompressionException) but will be converted to delegators at
a later date.