Removed populateImageMetadata from abstract layer
The empty implementation of populateImageMetadata has been removed from
AbstractMetadata. This is to encourage its implementation with each new
concrete metadata type. This information must always be known, and its
generation will be different for each format, so there is no reason to
have an empty implementation in the abstract layer.