Refactored SCIFIO Gateway class
SCIFIO.java now implements the org.scijava.Gateway interface, which
provides conveience methods for creating services, and raises exceptions
when those services are not found.

This allows for consolidation of null Service checking and simplifies
the accessor methods in the SCIFIO class itself.

Also, for consistency the accessor methods were renamed to drop the
plural form (e.g. formats() -> format()).

Note that to take advantage of the Gateway interface, scijava-common was
upgraded to 1.2.0, and pom-scijava to 1.34.