Added MetaTable for image/dataset metadata
Previously, Hashtables were used to track key, value pairs of metadata
at both the Image and Dataset level. These were confusingly named
"ImageMetadata" and "DatasetMetadata".

We have unified the concepts with a single "MetaTable" interface and
default implementation, and a "HasMetaTable" interface which both
ImageMetadata and Metadata implement.

Also the API for interacint with these tables has been simplified.
Updating them now requires a call to getTable() of the appropriate scope
(that is, from an Image or DatasetMetadata object). Then the table can
be manually populated or queried.