Converted QTFormat
Converted the LegacyQTReader/Writer and NativeQTReader/Writer from
loci.formats to LegacyQTFormat and NativeQTFormat, respectively, in
ome.scifio.formats.

Removed the DelegateReader class, instead adding an enabled/disabled
logic to the Format interface.

The NativeQTFormat has a higher priority than LegacyQTFormat so it will
always be chosen first. To switch to LegacyQTFormat, simply disable the
NativeQTFormat.