Improved Metadata support in SCIFIOFormatReader
Instead of always creating new Metadata, SCIFIOFormatReader now checks
for attached Metadata and uses that if available.

To support this functionality, added a protected setupMetadata method
so that individual Reader implementations can populate their metadata as
needed (e.g. attaching a MetadataStore to be populated).

This protected method was necessary because the Reader (and thus the
Metadata) is closed during setId calls.