Converted TIFFFormat
The loci.formats TiffReader and TiffWriter were consolidated to SCIFIO
components in ome.scifio as TIFFFormat. The supporting TiffSaver was
also copied to SCIFIO.

As TIFFFormat extended BaseTiffReader, it was also marked deprecated
(although its logic is restricted to a BaseTIFFParser in the new
TIFFFormat).

The TiffDelegateReader was also marked deprecated although it should
still function as normal.