Updated ZipHandle length calculation logic
ZipHandle was only doing a long length calculation when length was < 0,
but 0 was the default value so it was never being calculated.
setFile now sets the length to -1, so it is updated upon calling
resetStream.

Also, a stream was needed for the long length calculation, but the
length was being checked when the attached stream would always be null.
The length check now happens after setting the stream, and if the stream
is consumed to compute length, it is reset again.