Created SCIFIO Cell package
The SCIFIOCell series of classes facilitate creating an imgplus
(SCIFIOImgPlus) that uses the CellImg principles of dynamically loading
pixel information as it's requested.

Cells allow an image to be broken down into manageable bits, so that
images too large for memory can still be viewed and operated on.

The SCIFIOCellCache performs actual caching operatings, allowing
operations on cells to persist after the cell falls out of scope. It
uses an in-memory weak reference cache for fast memory action, and an
Ehcache back-end for its disk caching capabilities (and general memory
management).

This change also involves migrating ImgOpener and ImgSaver from Imglib
to SCIFIO (as these classes use SCIFIO to perform their IO operations,
and SCIFIO uses Imglib).