Package rename: ome.scifio to io.scif
The ome.scifio package has been removed in favor of io.scif. This is the
first stage in separating the scifio-devel component to its own product.
It also avoids package clashing with Bio-Formats 4.5.x which contains a
scifio-devel component under the ome.scifio namespace, but not the new
SCIFIO API.