Added getWriterByExtension method
Added a method to FormatService to directly get a writer for a given
id, based purely on suffix matching.

This allows bypassing of the Checker logic, which may not be applicable
when an Image source is not available for opening - because it has not
been written yet.

Only returns non-DefaultWriter instances.