Eliminate SNAPSHOT dependency on parent POM
This change has substantial ramifications:

1) It reiterates some configuration from what was previously the
   "pom-scifio" project parent.

2) It changes the transitional Bio-Formats artifact dependencies from
   4.5-SNAPSHOT to 4.4.8. This pins all of core SCIFIO's dependencies to
   stable release versions. But it also creates a slight version
   mismatch for the multi-module reactor as a whole (e.g., for
   components/bio-formats). Fortunately, Maven dependency versions are
   not fixed in stone unless you explicitly do it (by enclosing the
   version in "[]" brackets), so surprisingly, everything still compiles
   with passing tests.

The eventual goal is to migrate Bio-Formats-specific components into a
scifio/bio-formats fork, in preparation for later contribution back to
Bio-Formats upstream. Non-Bio-Formats-specific project forks and stubs
(i.e., jai_imageio and lwf-stubs) will live in the scifio project space
on GitHub as separate Git repositories, versioned separately (they are
almost never updated).

But for the moment, this structure works to build and deploy the code...