Corrected ColorTable population for PlanarImgs
When reading plane data for PlanarImgs, the reader metadata was
mistakenly being checked for a color table.

This was always false, because the reader filter was always wrapped with
a ChannelSeparator that prevented the underlying Metadata from being
examined.

Also, because the plane was just read, its color table was readily
available and is now used instead.