Fixed TIFF format priorities
MinimalTIFF and TIFFFormat were being placed at the same priority, which
was causing MinimalTIFF to be sorted ahead of TIFF because M<T.

Corrected this by adding explicit constants in each class that uses
priority more than once, with TIFF/OMETIFF each adding 1 to MinimialTIFF
priority. TIFFJAI is now MinimalTIFF - 1 (instead of TIFF - 1, which
would be equal to MinimalTIFF).