More robust default cell size determination
Split out the cell size determination to two methods in
SCIFIOCellImgFactory - one that takes a Reader only and uses the
optimalTileHeight/Width, and one that allows a starting tile
height/width to be specified.

Tiles are expanded, horizontally first and then vertically, to determine
an optimal cell size. Priority is given to spanning the entire width of
the image, and then the entire height of the image, even if it can't be
perfectly tiled.

Tiles are limited to 2MB.