Moved MinMaxFilter to io.scif
This was in an ome.xml package because of the use of a MinMaxStore.
However, there doesn't seem to be any compelling need for it in the code
base, and the filter is needed in io.scif for ImgOpener. Thus it was
moved.

An OME-XML subclass could be added later when needed that provides a
storage mechanism.