Added ImgFactoryHeuristic option
Created the ImgFactoryHeuristic interface for allowing custom logic when
selecting an ImgFactory. This can be specified in ImgOptions via the
setImgFactoryHeuristic method.

ImgOpener will only use an ImgFactoryHeuristic if a signature without an
ImgFactory is invoked. If no ImgFactoryHeuristic is specified, the
DefaultImgFactoryHeuristic will be used.

DefaultImgFactoryHeuristic is a default implementation that checks a
provided list of ImgModes in order. It uses a 2GB limit for ArrayImg
datasets or PlanarImg planes, and currently provides
SCIFIOCellImgFactories if the dataset doesn't fit in memory.