FakeFormat: use a default name if none is given
We assume that "none is given" when the first token contains an equals
sign. It is highly unlikely that the desired behavior is a name with an
equals sign. Rather, it is more likely that the first token constitutes
a key/value pair like any other, and the dataset name is not important.

Noticed by Johannes Schindelin.