Use hash to determine dirty status
Removed dirty flag from SCIFIOCell. Instead, comparing hashes of the
underlying data is used to determine dirty status.

This removes the ability to mark a cell as dirty without computing its
hash, but allows equality tests to include the underlying data.

Note that the dirty() method does not re-compute a hash of the
underlying data; that must be done via the update() method.