ImageMetadata: remove obsolete sub-dims metadata
The Bio-Formats sub-C dimension logic was a workaround for the fact that
Bio-Formats was limited to five dimensions. SCIFIO has no such
limitation, and we should not propagate this complicated and obsolete
workaround any further.

It also appears that the channel subdimension metadata was misused in
some places (in particular, in TextFormat).

Much further work still remains to be done to fully eliminate the
ZCT-specific case logic in many places. Ultimately we should eliminate
the "dimension order" string, since it is 5D-specific, in favor of
simply asking for the list of dimensional axes from the ImageMetadata.

We also need to clean up the "interleaved" and "RGB" logic, since that
is very channel-specific, and Bio-Formats has some nasty hacks (in
particular: IFormatReader#isInterleaved(int)) in order to support cases
where *some* of the sub-C dimensions are interleaved, but others aren't.
The correct answer, which is what SCIFIO should ultimately do, is to
return the list of dimensions in the actual order they are in the file,
along with some sort of hint as to how many dimensional axes are
embedded per plane (typically, but not necessarily, either 2 or 3).
This number has traditionally been called "RGBChannelCount" but that is
a horrible name which we should change.

Note that as things stand, ICS is untested and probably has bugs. But
solving them will be worth the pain of making ICS truly map properly
into the SCIFIO Metadata API rather than shoehorning it in to an
inadequate API as Bio-Formats does.