Make SCIFIO components extend AbstractSCIFIOComponent
These classes all extended AbstractHasSCIFIO before, but we also want
them all to have easy access to a LogService, too (and potentially other
things in the future, which can get unioned in to the SCIFIOComponent
interface as needed).

This change also makes a couple of corresponding interfaces (Filter,
Format and Translator) actually extend the SCIFIOComponent interface as
expected, rather than only having the abstract superclasses do so in
practice (by extending AbstractSCIFIOComponent).