README.md: update user info section
It no longer mentions Bio-Formats explicitly, because that is mentioned
in the developer section now. (It was also slightly misleading because
stock Bio-Formats does not yet support SCIFIO formats out of the box;
you need SCIFIO's temporary Bio-Formats fork for that.)

The user section also mentions Fiji, because Fiji now ships with ImageJ2
and can "Switch to Modern Mode" to utilize SCIFIO. Further, we plan for
Fiji's HandleExtraFileTypes to lean on SCIFIO much more heavily soon,
even when in ImageJ 1.x "legacy" mode.