IO: throw ImgIOException rather than logging
The main reason is that we are no longer using static SLF4J logging, so
it becomes more difficult for static utility classes to log error
messages. But it is also useful to propagate the ImgIOException to the
caller ito communicate more information about what went wrong in a
programmatic way.