Fix up @author tags
Use real names, and remove email addresses. The formatting was
inconsistent, and the same email address was not always used for the
same author. The email addresses are recorded in the Git commits, as
well as in the Maven POMs, so putting them in the @author tags too is
both redundant and error-prone.