Avoid overspecifying plugin type
This makes it easier to understand (for both humans and machines) that a
particular plugin belongs to a certain class of plugins. In particular,
ImageJ's System Information plugin can more easily group them together
as expected, without reporting many single-plugin types.

This also has the advantage of not loading the plugin classes themselves
until they actually need to be instantiated; SezPoz only needs to load
the plugin *types*, not the plugin *implementations*.