DefaultFormatService: remove unwanted dependency
The FormatService itself must not depend on implementation details of
its Format plugins, even if they are core and expected to be present.
After all, someone could repackage the library to remove unneeded core
formats. Conversely, and more importantly, there could be additional
Format plugins on the classpath which have additional dependencies, and
it must not be necessary to add those Formats' needed services to the
FormatService itself, or else we do not have a truly extensible system.

This reverts commit 44d9844a15340fbdbfa7e6a1afa7a442c9f3f9db.