AbstractImageMetadata: remove redundant assignment
There is no need to assign numeric primitive instance fields to 0, as
they are already set to zero by default. We already rely on this
behavior for booleans (they are all set to false by default), so we may
as well be consistent and rely on it for ints, too.