N-dimensional refactoring
The SCIFIO Metadata model is now N-dimensional, instead of the fixed 5-D
as implemented by Bio-Formats.

This allows an arbitrary number of axes per image. Ordering of axes is
still indicated by the order in the metadata, but valid orders are no
longer restricted to "XY....".

A new attribute, "planar dimensions", indicates the number of dimensions
that define a single image plane. For example, if given an image with
X=256, Y=256, C=3 and 2 planar dimensions, the image is 3 256x256
planes. If there are 3 planar dimensions for the same dataset, then the
C is the RGB channel. If there were 3 planar dimensions and the order
was CXY instead, it would indicate an RGB interleaved dataset.

There is a semi-universal assumption that the X and Y axes are still
present as planar axes. Removal of this assumption would require an
additional refactoring.

This single commit makes all Metadata changes and updates all utility
classes and Formats to comply to the new N-D model.