Fix up axis API and variable references
Mostly for historical reasons, there was substantial conflation between
CalibratedAxis and AxisType. We now refer to CalibratedAxis objects as
'axis', and arrays/collections of them as 'axes'. We refer to AxisType
objects as 'axisType', and arrays/collections of them as 'axisTypes'.
The API of various interfaces has also been standardized to use the same
nomenclature based on which arguments it takes and what it really does.