FormatTools: add getScale convenience method
This method uses the CalibratedAxis#averageScale(int, int) method to
return the average scale over the length of the axis. For linear axes
(which will be by far the most common), this will simply be the slope of
the axis (i.e., its "calibration value"). However, for nonlinear axes
this will average the scale over the axis, which is better than nothing.