Metadata: Standardize planar/interleaved counts
Changed the implicit interleaving of axes (reflected by isInterleaved)
to a specified interleaved axis count (set/getInterleavedAxisCount
methods).

There are now effectively three "classes" of axes - planar axes are the
first [0-planarAxiscount) axes, with the remaining being non-planar
axes.

Interleaved axes are a subset of planar axes (interleavedAxisCount <
planarAxisCount).

In this commit we officially establish Axes.X and Axes.Y as the
"canonical" planar axes. For consistency, planarAxiscount and
interleavedAxisCount are not required to be set, but will be
automatically detected based on the X and Y axis positions.