ImgOpener: don't overwrite separated axes
When looking for Axes to separate on, the unwrapped Metadata is now used
instead of the wrapped Metadata. This prevents, for example, automatic
separation of an axis by PlaneSeparator to be undone (because the
dataset looks like XY, then the separated axes are reset to be CXY
because ImgOpener decides to split on nothing).