TIFFFormat: improve imagej comment logic
When there were ImageJ comments specifying a channel count, this value
was only being respected if the image was known to be RGB (and the
previous channels length was overwritten).

The intended behavior was that the channels specified in the ImageJ
comment are never to be treated as RGB channels. So, the logic was
updated such that if the image is known to be RGB (e.g. RGB channels
defined in the IFD parsing) then the ImageJ channels will be added under
a different label (as a non-planar axis). If the image is not RGB, it's
fine to use the Axes.CHANNEL label.

fixed #80