FormatTools: assertId checks Objects
There was no reason to assertId on Strings only. This is just a null
object check, and since we now use RandomAccessInputStreams
predominantly, it makes sense to take an Object instead - allowing
either a String or RAIS.

Eventually this logic will go away when we refactor for IDs anyway. This
prevents inevitable NPEs when the RAIS is null, and it should be null,
but RAIS.getFileName() was being checked.