ImgSaver: fix composite channel count handling
Corrected how ImgSaver was using its source img's compositeChnanelCount
(which it wasn't using at all).

It now populates ImageMetadata using its source img, then if the
compositeChannelCount is > 1, the Channel axis length is updated (and
split into two axes if not all channels are to be composited) and
updates the planar axis count.

The actual saving algorithm then also needed to be updated to
consolidate all composite channels for each plane and write them out as
a single SCIFIO plane.

The SCIFIOMetadataTools API was modified slightly to facilitate these
changes.