IO: fix javadoc links -- sort of...
The links referred to the wrong method signatures. However, even the
"correct" method signatures substitute RealType for T, which is
technically incorrect because T extends both RealType and NativeType.
The Eclipse compiler still warns about it.

The weird thing, though, is that in some other cases with RealType
substituted for T, the Eclipse compiler is fine with it...