ImageIOFormat: populate img on translation
The ImageIOFormat translator was setting some metadata, but not
populating the actual BufferedImage of the destination Metadata. Now,
the source Metadata is used to generate a blank BufferedImage, which
provides information that was previously being missed (e.g.
RGB/interleaved information).

This will fix issues with ImageIO formats, such as JPEGFormat.