NativeQTFormat: use lazy compression detection
The NativeQTFormat writer had a constructor that attempts to use its
context, which is not set during invocation of the constructor.

The context was being used to determine supported compression types.
Instead, now we simply override getCompressionTypes() and if the
compression types are null, we can safely use a context to make the
check.