DefaultNIOService: avoid initialize() method
In practice we have found that relying on Service#initialize method can
have unintuitive consequences. At minimum, it is rather easy (and
natural) to put expensive initialization logic into the initialize
method, slowing down the startup of a context.

We have started migrating more and more toward a lazy initialization
pattern. To encourage this, the initialize method itself may be removed
in scijava-common 3.0.0. Since DefaultNIOService was the only SCIFIO
service making use of it, let's be proactive here and stop using it now.