FormatTools: add nextPlaneIndex methods
Added a set of N-D convenience methods. This method takes an arbitrary
length position array and gets the next valid plane index, per
positionToRaster, if it exists. It also updates the position array by
reference, and resets indices to min values when appropriate.

This method can be used to avoid recursive descent when iterating over
an arbitrary number of dimensions. It can handle cropping, but not
arbitrary indices in a range.