Add ImageViewer classes
Ported the Bio-Formats ImageViewer to SCIFIO. Also included are the
PreviewPane, ComboFileFilter, FormatFileFilter and a GUIService
(formerly GUITools) that are used in the ImageViewer and related
classes.

The original ImageViewer was built as a basic app for opening and saving
images. This functionality has been restricted by removing the menu bar,
so commands can be invoked programmatically only.

The primary purpose of this class right now is to provide basic viewing
functionality, e.g. for command line tools.

Use of ImageViewer outside of command line tools is not explicitly
guaranteed right now. The entire philosophy of maintaining our own
viewer is currently up for debate - an alternative being simply using
IJ2.