AWTImageTools: add openImage(..byte[]..) signature
Added another openImage signature that takes a byte[]. Although the
plane.getBytes method is sufficient for in-place modification of the
byte[], there may be cases where a copy of the data is made and
modified.

Thus, this method can now explicitly specify the byte[] that will be
used to open the BufferedImage.

If no byte[] is specified, plane.getBytes() will be used.