Add io.scif.refs package
Adds a RefManagerService for managing Reference implementations. When
there are references being maintained, a dedicated thread will poll a
ReferenceQueue in which these references were enqueued. When they are
dequeued, their cleanup method will be invoked.

This service is pared with the CleaningRef interface, which defines the
cleanup method. When implementing this interface, a type of Reference
should be extended, which will determine whne the CleaningRef will be
enqueued.

Also provided is a new plugin type - RefProvider. These are intended to
be singleton plugins bound to specific CleaningRef implementations,
responsible for checking compatibility and creating CleaningRef
instances.