IO utility class: dispose context
Added a new ImgPlusCtxCleaningProvider. This RefProvider creates
references to ImgPlus instances. When the ImgPluses are garbage
collected and its reference dequeued, if there are no remaining live
ImgPluses, the context used to create this ImgPlus is disposed.

This helps to mitigate contexts that are eternal because they were
created anonymously, e.g. by creating a new ImgOpener in the IO class.

Thus the IO class now registers its ImgPlus instances with the
RefManagerService.