Translator refactor: remove Metadata.CNAME
The CNAME constant has been removed from io.scif.Metadata and all
affected subclasses. Note that some subclasses did not define a CNAME,
which highlights the confusion and complexity of using the static
constant.

As a whole, the @Attrs that were previously used for Translator
discovery have been converted to API methods. The TranslatorService has
been updated to use the new API.