Filters: remove @Attr annotation usage
The Filter implementation was heavily relying on the @Attr addition to
the @Plugin annotation to define the filter target, priority and default
enabled status. This was extremely verbose and confusing, and
complicated the logic of Filter enabling and construction.

By converting the @Attr to API use, via new enabledDefault() and
target() methods, and overriding getPriority to return a fixed value,
the Filter logic has been greatly simplified.

Updated the FilterTest and filter usage as needed.