MetadataWrapper: remove plugin annotations
All ReaderFilters define the Metadata they are bound to in their
constructor. There is really no reason to be trying to dynamically
discover these. The Filter should be the unit of discovery, with any
accessory classes bound by their use.

This also removes the messy @Attr use in several MetadataWrapper
implementations.

The AbstractReaderFilter#setParent method has been updated to use the
known metadata class (if available) to create a new MetadataWrapper and
wrap the parent metadata. This logic is significantly simplified.