ImgSaver: remove type param from ImgPlus
Refactored all the typed ImgPlus signatures from <T extends RealType &
NativeType> to wildcards.

The type of the ImgPlus doesn't really matter for the writing algorithm,
and given that ImgOpener can't always guarantee the returned ImgPlus
type, requiring a generic parameter restricts the potential for
pipelining between ImgOpener and ImgSaver.

Had to refactor the ImgUtilityService#makeType(RealType) method to
accept an Object instead. This parameter is being type checked anyway in
the actual implementation, and given that Img's are generic typed on
<T>, and thus Objects and not just RealTypes, it makes sense to take
Objects and not RealTypes.