IO: actually catch exceptions
The javadoc for the IO utility class indicates that all exceptions
should be caught automatically, with nulls returned if there is a
problem, but many signatures did still throw exceptions.

The javadoc intention was the intended function of the IO class, thus
it has been updated to catch and log errors.