TIFFFormat: always save channels as RGB
The TIFFFormat translator was adding an IFD for each plane, and also
setting metadata for samples per pixel to be Axes.CHANNEL in length.

Thus if Axes.CHANNEL was not part of the source Metadata's planar axis
list, it was being counted twice. This resulted in a duplication of the
Channel axis (once as channels, once as time points).

This caused problems e.g. when converting the type of an image in IJ2,
it's always created as a non-RGB image, so channels are split to
non-planar axes. Then when saving the new typed image as a .tif, the
plane count could be off by a factor of the channel count, and you'd get
an AIOOB.

This should now be fixed, by ALWAYS treating Axes.CHANNEL as RGB
channels for tiff. This seemd more appropriate than treating them as
time points, but there may be a better solution.

Closes #111.